chore: rename project module
