+++ olake-workers/k8s/utils/filesystem/filesystem.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"olake-ui/olake-workers/k8s/shared"
+	"github.com/datazip-inc/olake-ui/olake-workers/k8s/shared"
 )
 
 // Helper handles file system operations for job configuration
